requirements.txt
python-dotenv==1.0.1
transformers==4.41.1
tensorflow==2.16.1
langchain==0.2.1
tf-keras==2.16.0
pillow == 10.3.0
torch==2.2.0
accelerate ==0.30.1
ollama
streamlit